mvoe persistance methods to main keeper file, consolidate tests
